move tests into the shared suite
