--- playground/js-sourcemap/__tests__/downleveled-optimized-dep-transformed-again/root/index.html
@@ -1,6 +1,0 @@
-<div class="wrapper">
-  <h1>Downleveled Optimized Dep Transformed Again</h1>
-</div>
-
-<script type="module" src="./optimized-class-field-import-babel.js"></script>
-<script type="module" src="./optimized-class-field-import-oxc.js"></script>
