--- playground/js-sourcemap/__tests__/downleveled-optimized-dep-transformed-again/serve.ts
@@ -1,14 +1,0 @@
-import fs from 'node:fs/promises'
-import path from 'node:path'
-import { rootDir, startDefaultServe, viteServer } from '~utils'
-
-export async function preServe() {
-  await fs.cp(path.resolve(import.meta.dirname, 'root'), rootDir, {
-    recursive: true,
-  })
-}
-
-export async function serve() {
-  await startDefaultServe()
-  return viteServer
-}
