+++ playground/js-sourcemap/dep-class-field-sourcemap-babel/index.js
