+++ playground/js-sourcemap/dep-class-field-sourcemap-babel/package.json
