+++ playground/js-sourcemap/index.html
@@ -13,3 +13,5 @@
 <script type="module" src="./with-define-object.ts"></script>
 <script type="module" src="./malicious-import.js"></script>
 <script type="module" src="./optimized-malicious-import.js"></script>
+<script type="module" src="./optimized-class-field-import-babel.js"></script>
+<script type="module" src="./optimized-class-field-import-oxc.js"></script>
