+++ playground/js-sourcemap/optimized-class-field-import-babel.js
