+++ playground/js-sourcemap/vite.config.js
@@ -1,3 +1,7 @@
+import { fileURLToPath } from 'node:url'
+import babel from '@babel/core'
+import MagicString from 'magic-string'
+import { Visitor, parseSync } from 'oxc-parser'
 import { defineConfig } from 'vite'
 import transformFooWithInlineSourceMap from './foo-with-sourcemap-plugin'
 import { transformZooWithSourcemapPlugin } from './zoo-with-sourcemap-plugin'
@@ -6,9 +10,81 @@ export default defineConfig({
   plugins: [
     transformFooWithInlineSourceMap(),
     transformZooWithSourcemapPlugin(),
+    {
+      name: 'test-babel-transformed-optimized-dep',
+      async transform(code, id) {
+        if (
+          !id.includes('/deps/') ||
+          !id.includes('@vitejs_test-dep-class-field-sourcemap-babel.js')
+        ) {
+          return
+        }
+
+        return babel.transformAsync(code, {
+          filename: id,
+          configFile: false,
+          babelrc: false,
+          ast: false,
+          sourceMaps: true,
+          plugins: [() => ({ visitor: { Program() {} } })],
+        })
+      },
+    },
+    {
+      name: 'test-oxc-transformed-optimized-dep',
+      transform(code, id) {
+        if (
+          !id.includes('/deps/') ||
+          !id.includes('@vitejs_test-dep-class-field-sourcemap-oxc.js')
+        ) {
+          return
+        }
+
+        const parsed = parseSync(id, code, { range: true })
+        const s = new MagicString(code)
+        const visitor = new Visitor({
+          Identifier(node) {
+            s.overwrite(node.start, node.end, `${node.name}$$$`, {
+              storeName: true,
+            })
+          },
+        })
+        visitor.visit(parsed.program)
+
+        return {
+          code: s.toString(),
+          map: s.generateMap({
+            hires: true,
+            source: id,
+            includeContent: true,
+          }),
+        }
+      },
+    },
   ],
+  resolve: {
+    alias: {
+      '@vitejs/test-dep-class-field-sourcemap-babel': fileURLToPath(
+        new URL('./dep-class-field-sourcemap-babel/index.js', import.meta.url),
+      ),
+      '@vitejs/test-dep-class-field-sourcemap-oxc': fileURLToPath(
+        new URL('./dep-class-field-sourcemap-oxc/index.js', import.meta.url),
+      ),
+    },
+  },
   optimizeDeps: {
     exclude: ['@vitejs/test-dep-malicious-sourcemap'],
+    include: [
+      '@vitejs/test-dep-class-field-sourcemap-babel',
+      '@vitejs/test-dep-class-field-sourcemap-oxc',
+    ],
+    // Force class fields to be downleveled so the optimized deps get the
+    // constructor parameter rename shape some of these sourcemap tests assert on.
+    rolldownOptions: {
+      transform: {
+        target: 'es2015',
+      },
+    },
   },
   build: {
     sourcemap: true,
